1.  Add PanelRoute table to database which tracks what routes and panel's output LED is assigned to 2.  More work on converting UDPTester to use json/ControlServer

+++ ControlServer/ControllerManager.cpp
@@ -122,7 +122,7 @@ void ControllerManager::getConnectedInfo(int serialNumber, int &version, Control
     }
 }
 
-void ControllerManager::controllerResetting(int serialNumber)
+void ControllerManager::controllerResetting(long serialNumber)
 {
     qDebug(QString("Controller %1 restarting!!!!!!!").arg(serialNumber).toLatin1());
